Move message_waiter to DiscordWaiter.py
