+++ customs/Utility.py
@@ -1,11 +1,6 @@
 import logging
-from typing import Callable, Tuple, Type, Optional
+from typing import Tuple, Type, Optional
 from urllib.parse import quote_plus as qp, urlparse
-
-import discord
-from discord.ext import commands
-
-import config
 
 __all__ = ["Utility"]
 
@@ -60,23 +55,3 @@ class Utility:
     @staticmethod
     def is_an_url(url: str) -> bool:
         return urlparse(url).netloc != ""
-
-    @staticmethod
-    def message_waiter(ctx: commands.Context) -> Callable[[discord.Message], bool]:
-        """
-        Message waiter to use with Client.wait_for("message", ...).
-
-        Usages:
-            message: discord.Message = await bot.wait_for("message", check=wait_for)
-
-        :param ctx: Current context.
-        :return: The waiter function.
-        """
-
-        def message_checker(message: discord.Message) -> bool:
-            return (
-                message.author.id == ctx.author.id
-                and ctx.channel.id == message.channel.id
-            )
-
-        return message_checker
